refactor: details in commit description
- add simple custom_id randomization
- extract drag handler svg to a separate file
- remove old css
- change directory name to dnd
