+++ components-sdk/src/components/Section.module.css
@@ -7,9 +7,5 @@
     position: relative;
   }
 
-  .left {
-    /*padding-left: calc(1rem + 5px);*/
-  }
-
 }
